moves configure monitoring and creat project finished adapters to monitoring package
